refactor: Class로부터 controller 오브젝트 생성 책임 분리
